requirements.txt
Django~=5.1.4
djangorestframework-simplejwt~=5.3.1
djangorestframework~=3.15.2